chore(utils): move promise timeout in infrastructure
